refactor(users): change inputsConfig import dir
